+++ web_debranding/__openerp__.py
@@ -1,6 +1,6 @@
 {
     'name': "Backend debranding",
-    'version': '1.0.5',
+    'version': '1.0.6',
     'author': 'IT-Projects LLC, Ivan Yelizariev',
     'license': 'LGPL-3',
     'category': 'Debranding',